+++ lib/src/inverted_index/in_memory_index.dart
@@ -5,17 +5,16 @@
 import 'package:text_indexing/src/_index.dart';
 
 /// The InMemoryIndex is an implementation of the [InvertedIndex] interface
-/// that mixes in [InvertedIndexMixin] and [InMemoryIndexMixin].
+/// that extends [InMemoryIndexBase].
 ///
 /// The InMemoryIndex is intended for fast indexing of a smaller corpus using
 /// in-memory dictionary, k-gram and postings hashmaps.
-class InMemoryIndex
-    with InMemoryIndexMixin, InvertedIndexMixin
-    implements InvertedIndex {
-  //
-
+class InMemoryIndex extends InMemoryIndexBase {
   @override
   final int k;
+
+  @override
+  final TokenizingStrategy strategy;
 
   @override
   final NGramRange nGramRange;
@@ -29,45 +28,80 @@ class InMemoryIndex
   @override
   late PostingsMap postings;
 
-  /// Instantiates a [InMemoryIndex] instance:
-  /// - [tokenizer] is the [TextTokenizer] used to tokenize text for the index;
-  /// - [k] is the length of k-gram entries in the k-gram index;
-  /// - [nGramRange] is the range of N-gram lengths to generate;
-  /// - [zones] is a hashmap of zone names to their relative weight in the
-  ///   index;
-  /// - [dictionary] is the in-memory term dictionary for the indexer. Pass a
-  ///   [DftMap] instance at instantiation, otherwise an empty [DftMap]
-  ///   will be initialized;
-  /// - [kGramIndex] is the in-memory [KGramsMap] for the index. Pass a
-  ///   [KGramsMap] instance at instantiation, otherwise an empty [KGramsMap]
-  ///   will be initialized; and
-  /// - [postings] is the in-memory postings hashmap for the indexer. Pass a
-  ///   [PostingsMap] instance at instantiation, otherwise an empty [PostingsMap]
-  ///   will be initialized.
-  InMemoryIndex(
-      {required this.tokenizer,
-      DftMap? dictionary,
-      PostingsMap? postings,
-      KGramsMap? kGramIndex,
-      this.k = 2,
-      this.nGramRange = const NGramRange(1, 2),
-      this.zones = const <String, double>{}}) {
-    this.dictionary = dictionary ?? {};
-    this.postings = postings ?? {};
-    this.kGramIndex = kGramIndex ?? {};
-  }
+  @override
+  late KeywordPostingsMap keywordPostings;
 
   @override
   final TextTokenizer tokenizer;
 
   @override
+  final KeywordExtractor keywordExtractor;
+
+  @override
   final ZoneWeightMap zones;
+
+  /// Instantiates a [InMemoryIndex] instance.
+  /// - [tokenizer] is the [TextTokenizer] used to tokenize text for the index;
+  /// - [keywordExtractor] is a splitter function that returns an ordered
+  ///   collection of keyword phrasesfrom text.
+  /// - [k] is the length of k-gram entries in the k-gram index.
+  /// - [nGramRange] is the range of N-gram lengths to generate.
+  /// - [strategy] is the tokenizing strategy to use when tokenizing documents
+  ///   for indexing.
+  /// - [zones] is a hashmap of zone names to their relative weight in the
+  ///   index.
+  /// - [dictionary] is the in-memory term dictionary for the indexer. Pass a
+  ///   [DftMap] instance at instantiation, otherwise an empty [DftMap]
+  ///   will be initialized.
+  /// - [kGramIndex] is the in-memory [KGramsMap] for the index. Pass a
+  ///   [KGramsMap] instance at instantiation, otherwise an empty [KGramsMap]
+  ///   will be initialized.
+  /// - [postings] is the in-memory postings hashmap for the indexer. Pass a
+  ///   [PostingsMap] instance at instantiation, otherwise an empty [PostingsMap]
+  ///   will be initialized.
+  /// - [keywordPostings] is the in-memory postings hashmap for the indexer. Pass a
+  ///   [PostingsMap] instance at instantiation, otherwise an empty [PostingsMap]
+  ///   will be initialized.
+  InMemoryIndex(
+      {required this.tokenizer,
+      required this.keywordExtractor,
+      DftMap? dictionary,
+      PostingsMap? postings,
+      KeywordPostingsMap? keywordPostings,
+      KGramsMap? kGramIndex,
+      this.k = 2,
+      this.nGramRange = const NGramRange(1, 2),
+      this.strategy = TokenizingStrategy.terms,
+      this.zones = const <String, double>{}}) {
+    this.dictionary = dictionary ?? {};
+    this.postings = postings ?? {};
+    this.keywordPostings = keywordPostings ?? {};
+    this.kGramIndex = kGramIndex ?? {};
+  }
+
+  //
+}
+
+/// Base class implementation of [InvertedIndex] with [InMemoryIndexMixin] and
+/// [InvertedIndexMixin].
+///
+/// Provides a const default generative constructor for sub-classes.
+abstract class InMemoryIndexBase
+    with InMemoryIndexMixin, InvertedIndexMixin
+    implements InvertedIndex {
+  //
+
+  /// Const default generative constructor for sub-classes.
+  const InMemoryIndexBase();
+
+  //
 }
 
 /// A mixin class that implements [InvertedIndex]. The mixin exposes in-memory
-/// [dictionary] and [postings] fields that must be overriden. Provides
-/// implementation of the following methods for operations on the in-memory
-/// [postings] and [dictionary] maps:
+/// [dictionary] and [postings] fields that must be overriden.
+///
+/// Providesimplementation of the following methods for operations on the
+/// in-memory [postings], [keywordPostings] and [dictionary] maps:
 /// - [vocabularyLength] returns the number of entries in the in-memory
 ///   [dictionary].
 /// - [getDictionary] retrieves a [DftMap] for a collection of [Term]s from
@@ -78,6 +112,10 @@ class InMemoryIndex
 ///   in-memory [postings] hashmap;
 /// - [upsertPostings] inserts entries into the in-memory [postings] hashmap,
 ///   overwriting any existing entries;
+/// - [getKeywordPostings] retrieves [KeywordPostingsMap] for a collection of
+///   keywords from the in-memory [keywordPostings] hashmap;
+/// - [upsertKeywordPostings] inserts entries into the in-memory
+///   [keywordPostings] hashmap, overwriting any existing entries;
 /// - [getKGramIndex] retrieves entries for a collection of [Term]s from the
 ///   in-memory [kGramIndex] hashmap; and
 /// - [upsertKGramIndex] inserts entries into the in-memory [kGramIndex] hashmap,
@@ -94,8 +132,28 @@ abstract class InMemoryIndexMixin implements InvertedIndex {
   /// The in-memory [KGramsMap] for the index.
   KGramsMap get kGramIndex;
 
+  /// The in-memory keyword postings hashmap for the index.
+  KeywordPostingsMap get keywordPostings;
+
   @override
   Future<Ft> get vocabularyLength async => dictionary.length;
+
+  @override
+  Future<KeywordPostingsMap> getKeywordPostings(Iterable<Term> keywords) async {
+    final KeywordPostingsMap retVal = {};
+    for (final e in keywords) {
+      final entry = keywordPostings[e];
+      if (entry != null) {
+        retVal[e] = entry;
+      }
+    }
+    return retVal;
+  }
+
+  @override
+  Future<void> upsertKeywordPostings(KeywordPostingsMap values) async {
+    keywordPostings.addAll(values);
+  }
 
   @override
   Future<PostingsMap> getPostings(Iterable<Term> terms) async {