+++ lib/src/inverted_index/inverted_index.dart
@@ -6,111 +6,127 @@ import 'dart:math';
 import 'package:text_indexing/src/_index.dart';
 
 /// An interface that exposes methods for working with an inverted, positional
-/// zoned index on a collection of documents:
-/// - [tokenizer] is the [TextTokenizer] used to index the corpus terms;
-/// - [vocabularyLength] is the number of unique terms in the corpus;
+/// zoned index on a collection of documents.
+/// - [tokenizer] is the [TextTokenizer] used to index the corpus terms.
+/// - [keywordExtractor] is a splitter function that returns an ordered
+///   collection of keyword phrasesfrom text.
+/// - [vocabularyLength] is the number of unique terms in the corpus.
 /// - [zones] is a hashmap of zone names to their relative weight in the index.
-///   If [zones] is empty, all the `JSON` fields will be indexed;
-/// - [k] is the length of k-gram entries in the k-gram index;
+///   If [zones] is empty, all the `JSON` fields will be indexed.
+/// - [k] is the length of k-gram entries in the k-gram index.
 /// - [nGramRange] is the range of N-gram lengths to generate. The minimum
 ///   n-gram length is 1. If n-gram length is greater than 1, the
 ///   index vocabulary also contains n-grams up to [nGramRange].max long,
 ///   concatenated from consecutive terms. The index size is increased by a
-///   factor of [nGramRange].max;
+///   factor of [nGramRange].max.
+/// - [strategy] is the tokenizing strategy to use when tokenizing documents
+///   for indexing.
 /// - [getDictionary] Asynchronously retrieves a [DftMap] for a collection
-///   of [Term]s from a [DftMap] repository;
+///   of [Term]s from a [DftMap] repository.
 /// - [upsertDictionary ] inserts entries into a [DftMap] repository,
-///   overwriting any existing entries;
+///   overwriting any existing entries.
 /// - [getKGramIndex] Asynchronously retrieves a [KGramsMap] for a collection
-///   of [KGram]s from a [KGramsMap] repository;
+///   of [KGram]s from a [KGramsMap] repository.
 /// - [upsertKGramIndex ] inserts entries into a [KGramsMap] repository,
-///   overwriting any existing entries;
+///   overwriting any existing entries.
 /// - [getPostings] asynchronously retrieves [PostingsMap] for a collection
-///   of [Term]s from a [PostingsMap] repository;
+///   of [Term]s from a [PostingsMap] repository.
 /// - [upsertPostings] inserts entries into a [PostingsMap] repository,
-///   overwriting any existing entries;
+///   overwriting any existing entries.
 /// - [getTfIndex] returns hashmap of [Term] to [Ft] for a collection of
 ///   [Term]s, where [Ft] is the number of times each of the terms occurs in
-///   the `corpus`;
+///   the `corpus`.
 /// - [getFtdPostings] return a [FtdPostings] for a collection of [Term]s from
-///   the [PostingsMap], optionally filtered by minimum term frequency; and
+///   the [PostingsMap], optionally filtered by minimum term frequency.
 /// - [getIdFtIndex] returns a [IdFtIndex] for a collection of [Term]s from
 ///   the [DftMap].
 abstract class InvertedIndex {
   //
 
-  /// A factory constructor that returns an [InMemoryIndex] instance:
-  /// - [tokenizer] is the [TextTokenizer] used to tokenize text for the index;
-  /// - [k] is the length of k-gram entries in the k-gram index;
+  /// A factory constructor that returns an [InMemoryIndex] instance.
+  /// - [tokenizer] is the [TextTokenizer] used to tokenize text for the index.
+  /// - [keywordExtractor] is a splitter function that returns an ordered
+  ///   collection of keyword phrasesfrom text.
+  /// - [k] is the length of k-gram entries in the k-gram index.
+  /// - [strategy] is the tokenizing strategy to use when tokenizing documents
+  ///   for indexing.
+  /// - [nGramRange] is the range of N-gram lengths to generate.
   /// - [zones] is a hashmap of zone names to their relative weight in the
-  ///   index;
+  ///   index.
   /// - [dictionary] is the in-memory term dictionary for the indexer. Pass a
   ///   [DftMap] instance at instantiation, otherwise an empty [DftMap]
-  ///   will be initialized;
+  ///   will be initialized.
   /// - [kGramIndex] is the in-memory [KGramsMap] for the index. Pass a
   ///   [KGramsMap] instance at instantiation, otherwise an empty [KGramsMap]
-  ///   will be initialized; and
+  ///   will be initialized.
   /// - [postings] is the in-memory postings hashmap for the indexer. Pass a
   ///   [PostingsMap] instance at instantiation, otherwise an empty [PostingsMap]
   ///   will be initialized.
   factory InvertedIndex.inMemory(
           {required TextTokenizer tokenizer,
+          required KeywordExtractor keywordExtractor,
           Map<String, int>? dictionary,
           Map<String, Map<String, Map<String, List<int>>>>? postings,
+          KeywordPostingsMap? keywordPostings,
           Map<String, Set<String>>? kGramIndex,
           int k = 2,
           NGramRange nGramRange = const NGramRange(1, 2),
+          TokenizingStrategy strategy = TokenizingStrategy.terms,
           Map<String, double> zones = const <String, double>{}}) =>
       InMemoryIndex(
           tokenizer: tokenizer,
+          keywordExtractor: keywordExtractor,
           dictionary: dictionary,
           postings: postings,
+          keywordPostings: keywordPostings,
           kGramIndex: kGramIndex,
           k: k,
+          strategy: strategy,
           nGramRange: nGramRange,
           zones: zones);
 
-  /// /// A factory constructor that returns an [AsyncCallbackIndex] instance:
-  /// - [tokenizer] is the [TextTokenizer] used to tokenize text for the index;
-  /// - [k] is the length of k-gram entries in the k-gram index;
+  /// /// A factory constructor that returns an [AsyncCallbackIndex] instance.
+  /// - [tokenizer] is the [TextTokenizer] used to tokenize text for the index.
+  /// - [keywordExtractor] is a splitter function that returns an ordered
+  ///   collection of keyword phrasesfrom text.
+  /// - [k] is the length of k-gram entries in the k-gram index.
+  /// - [strategy] is the tokenizing strategy to use when tokenizing documents
+  ///   for indexing.
+  /// - [nGramRange] is the range of N-gram lengths to generate.
   /// - [zones] is a hashmap of zone names to their relative weight in the
-  ///   index;
+  ///   index.
   /// - [dictionaryLengthLoader] asynchronously retrieves the number of terms
-  ///   in the vocabulary (N);
+  ///   in the vocabulary (N).
   /// - [dictionaryLoader] asynchronously retrieves a [DftMap] for a
-  ///   vocabulary from a index repository;
+  ///   vocabulary from a index repository.
   /// - [dictionaryUpdater] is callback that passes a [DftMap] subset
-  ///    for persisting to a index repository;
+  ///    for persisting to a index repository.
   /// - [kGramIndexLoader] asynchronously retrieves a [KGramsMap] for a
-  ///   vocabulary from a index repository;
+  ///   vocabulary from a index repository.
   /// - [kGramIndexUpdater] is callback that passes a [KGramsMap] subset
-  ///    for persisting to a index repository;
+  ///    for persisting to a index repository.
   /// - [postingsLoader] asynchronously retrieves a [PostingsMap] for a vocabulary
-  ///   from a index repository; and
+  ///   from a index repository.
   /// - [postingsUpdater] passes a [PostingsMap] subset for persisting to a
   ///   index repository.
   factory InvertedIndex(
-          {required Future<Map<String, int>> Function([Iterable<String>?])
-              dictionaryLoader,
-          required Future<void> Function(Map<String, int>) dictionaryUpdater,
-          required Future<int> Function() dictionaryLengthLoader,
-          required Future<Map<String, Set<String>>> Function(
-                  [Iterable<String>?])
-              kGramIndexLoader,
-          required Future<void> Function(Map<String, Set<String>>)
-              kGramIndexUpdater,
-          required Future<Map<String, Map<String, Map<String, List<int>>>>>
-                  Function(Iterable<String>)
-              postingsLoader,
-          required Future<void> Function(
-                  Map<String, Map<String, Map<String, List<int>>>>)
-              postingsUpdater,
+          {required DftMapLoader dictionaryLoader,
+          required DftMapUpdater dictionaryUpdater,
+          required VocabularySize dictionaryLengthLoader,
+          required KGramsMapLoader kGramIndexLoader,
+          required KGramsMapUpdater kGramIndexUpdater,
+          required PostingsMapLoader postingsLoader,
+          required PostingsMapUpdater postingsUpdater,
+          required KeywordPostingsMapLoader keywordPostingsLoader,
+          required KeywordPostingsMapUpdater keywordPostingsUpdater,
           required TextTokenizer tokenizer,
+          required KeywordExtractor keywordExtractor,
           Map<String, int>? dictionary,
           Map<String, Map<String, Map<String, List<int>>>>? postings,
           Map<String, Set<String>>? kGramIndex,
           int k = 2,
           NGramRange nGramRange = const NGramRange(1, 2),
+          TokenizingStrategy strategy = TokenizingStrategy.terms,
           Map<String, double> zones = const <String, double>{}}) =>
       AsyncCallbackIndex(
           dictionaryLoader: dictionaryLoader,
@@ -120,13 +136,20 @@ abstract class InvertedIndex {
           kGramIndexUpdater: kGramIndexUpdater,
           postingsLoader: postingsLoader,
           postingsUpdater: postingsUpdater,
+          keywordPostingsLoader: keywordPostingsLoader,
+          keywordPostingsUpdater: keywordPostingsUpdater,
           tokenizer: tokenizer,
+          keywordExtractor: keywordExtractor,
           k: k,
           nGramRange: nGramRange,
+          strategy: strategy,
           zones: zones);
 
   /// The length of k-gram entries in the k-gram index.
   int get k;
+
+  ///The strategy to apply when splitting text to [Token]s.
+  TokenizingStrategy get strategy;
 
   /// The minimum and maximum length of n-grams in the index.
   NGramRange get nGramRange;
@@ -138,6 +161,10 @@ abstract class InvertedIndex {
 
   /// The text analyser that extracts tokens from text for the index.
   TextTokenizer get tokenizer;
+
+  /// A splitter function that returns an ordered collection of keyword phrases
+  /// from text.
+  KeywordExtractor get keywordExtractor;
 
   /// Returns the number of terms in the vocabulary (N).
   Future<Ft> get vocabularyLength;
@@ -199,6 +226,17 @@ abstract class InvertedIndex {
   /// Inserts [values] into a [PostingsMap] repository, overwriting them if they
   /// already exist.
   Future<void> upsertPostings(PostingsMap values);
+
+  /// Asynchronously retrieves [PostingsMapEntry] entities for the [terms] from a
+  /// [PostingsMap] repository.
+  ///
+  /// Used in `index-elimination`, to return a subset of the [PostingsMap]
+  /// where the key ([Term]) is in [keywords].
+  Future<KeywordPostingsMap> getKeywordPostings(Iterable<Term> keywords);
+
+  /// Inserts [values] into a [PostingsMap] repository, overwriting them if they
+  /// already exist.
+  Future<void> upsertKeywordPostings(KeywordPostingsMap values);
 }
 
 /// An abstract [InvertedIndex] base class that mixies in [InvertedIndexMixin].